LayoutTests/fast/images/reset-image-animation-expected.txt

Ensure the image animation is played in order after the animation is reset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.imageFrameIndex(image) is 1
PASS internals.imageFrameIndex(image) is 2
PASS successfullyParsed is true

TEST COMPLETE
